Experimentation/MnistGAN/requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
coverage==5.5
cycler==0.10.0
dataclasses==0.8
decorator==5.0.9
dill==0.3.4
easydict==1.9
flatbuffers==1.12
future==0.18.2
gans2==0.1
gast==0.3.3
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.5.0
importlib-metadata==4.5.0
iniconfig==1.1.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.1.0
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
overrides==3.1.0
packaging==20.9
pandas==1.1.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.0
pluggy==0.13.1
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.17.3
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
Pygments==2.9.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.10.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.1.2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-addons==0.13.0
tensorflow-datasets==1.3.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
tensorflow-metadata==1.0.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.47.0
traitlets==4.3.3
typeguard==2.12.1
typing-extensions==3.7.4.3
urllib3==1.26.5
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
